add map viewer of xoz and zoy planes

+++ 2010_hackathon/yang_jz/mapview/mapviewer.h
@@ -148,7 +148,7 @@ public:
 	
 	QString curFilePath;
 	
-	float scaleFactor;
+	int scaleFactor;
 	
 	static XMapView* m_show;
